Move scaffold-related components to its own package
